Move challenges into a subfolder
